bug 1428945: Move sections to feeder app
Sections are used to group feeds, and seem out-of-place in core.
